feat(connector): build RPC client and transaction prep on kit-native plumbing
- createSolanaClient now constructs rpc/rpcSubscriptions via createClient()
  + solanaRpcConnection from @solana/kit-plugin-rpc (new dependency)
- prepareTransaction estimates compute unit / resource limits through kit's
  estimateResourceLimitsFactory + estimateAndSetResourceLimitsFactory,
  un-stubbing computeUnitLimitMultiplier and computeUnitLimitReset
- lamportsToSol/solToLamports deprecated in favor of kit's exact fixed-point
  equivalents; unused GENESIS_HASH removed (internal-only)
- lib/utils/explorer-urls.ts merged into lib/kit/explorer.ts (single explorer
  module); public export paths unchanged via headless.ts
